Initial work on a new Webpack-based build.
It still uses Gulp as a harness (it makes it easier to produce the es5
and minified versions, as well as the bootstrap script), but relies on
webpack to assemble the main components bundle.  This does add a little
bit of overhead (in order to do the nice Webpack loading and code
isolation), but enables some much more complex, efficient build
patterns.