Update web component build version
Fixes #123

Also includes some incidental upgrades to other dependencies, and a new
option in the demo pages `load-components.js`.  Now, when viewing the
demo pages locally, you can add a query parameter (load-local=true) to
force the loader to load the local files instead of from the CDN. This
is useful when you're debugging locally, but not serving from
'localhost' (for example, when you're debugging IE11 inside of a VM on
MacOS).