Add support for serving with Content-Security-Policy (#201)
* Update byu-web-component-build version to fix #151

* emove eval() in byu-search

Fixes #151. `eval()` is considered unsafe is browser-side Javascript,
and is usually blocked on security-conscious sites. Our usage of eval()
can easily be replaced by looking up the function we want to invoke on
window, then calling it directly.

This commit also ensures that the provided callback is invoked with the
`byu-search` element as the `this` context, as is standard in DOM event
listeners.

* Run build