No overflow
By getting rid of the padding on the university-logo-wrapper, you get rid of the overflow on your pages. Also, a better way of formatting the footer would be providing the university-logo a width:100% and a max-width:450px, as well as giving the university-logo-wrapper a max-width of 1200px (or whatever the user desires) and a width of 100%. This will contain the size of the photo as well as supply a max width for both.  I also notice that the BYU Sample Theme site has more overflow issues due to the fact that the HTML and body have a width:100vm; and max-width: 100vm; if you switched this to only have a width:100%; it will get rid of unnecessary overflow.