Actually fix the safari issue
It seems like attribute changes on the host element do not trigger a
restyle.  However, class changes do. This changes the search hiding to
rely on a class toggle. The interesting thing is that just toggling the
class, even while having the styles depend on the attribute, triggers a
restyle and fixes the issue.

It looks like this is a bug in WebKit. Good luck getting it fixed.