Add CDN preload config (#383)
This will cause the CDN to add preload links to its responses.  This allows for better performance in the browser, as the browser can start fetching resources immediately instead of having to wait to parse and execute the responses before it finds out what it needs next.

This feature has been stable in the production CDN for some time, so it's time to move the configs from the main CDN config to this repo.