Improved fix for #382.
This fixes issues with referencing 'this' in the logic, as well as
making sure that the placeholder and title a11y helpers are exhaustive.

This has the side benefit of, in my opinion, making this part of the
code a bit easier to read than the original version (gee, I wonder who
could have written that!? I'm looking at you, past me!).